pyproject.toml
[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "cr-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai-slim>=0.8.1",
    "mcp>=1.0.0",
    "python-dotenv>=1.2.1",
    "openai>=1.0.0",
    "rich>=13.0.0",
    "argcomplete>=3.0.0",
    "pydantic-ai>=1.30.1",
    "gitpython>=3.1.45",
    "langchain>=1.1.3",
    "langchain-openai>=1.1.3",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-community>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "ipython>=8.37.0",
    "langchain-experimental>=0.4.1",
    "markdown-it-py>=3.0.0",
    "linkify-it-py>=2.0.0",
    "unidiff>=0.7.5",
]

[[tool.uv.index]]
name = "tencent"
url = "https://mirrors.cloud.tencent.com/pypi/simple/"
